+++ NEWS.md
@@ -7,6 +7,7 @@
 - Throw a `ContextError` for context values of an incorrect type
 - The `preamble` used in `render` and some `show` methods can now be specified using `set_preamble`
 - `render` now supports the `ignorestatus = true` keyword parameter
+- Emulation of Typst command line interface. `typst("compile input.typ output.pdf")`
 
 ### Bug Fixes